# Reduce some allocations in SslStream handshake.